+++ flopscope-client/src/flopscope/__init__.py
@@ -18,7 +18,7 @@ import builtins
 import struct
 from typing import Any
 
-__version__ = "0.4.1"
+__version__ = "0.4.2"
 
 # ---------------------------------------------------------------------------
 # Errors